+++ src/app/_components/chat/chatlist/ChatPreview.tsx
@@ -1,0 +1,5 @@
+const ChatPreview = () => {
+  return <div>ChatPreview</div>
+}
+
+export default ChatPreview
