+++ src/app/_components/chat/containers/ChatContainer.tsx
@@ -1,4 +1,4 @@
-import ChatList from '../components/ChatList'
+import ChatList from '../chatroom/ChatList'
 
 const ChatContainer = () => {
   return (
