+++ src/app/_components/chat/containers/ChatInputContainer.tsx
@@ -3,13 +3,22 @@
 import { ChangeEvent, useState } from 'react'
 import Icons from '@/_components/common/Icons'
 import { send } from '@/_ui/IconsPath'
+import useWebSocket from '@/_hooks/useWebSocket'
 import Input from '../../common/Input'
 
 const ChatInputContainer = () => {
   const [inputVal, setInputVal] = useState<string | null>(null)
+  const { sendMessage } = useWebSocket()
 
   const onChangeInput = (e: ChangeEvent<HTMLInputElement>) => {
     setInputVal((prev) => e.target.value)
+  }
+
+  const handleSendMessage = () => {
+    if (inputVal) {
+      sendMessage(1, inputVal) // roomId 일단 1로.
+      setInputVal('')
+    }
   }
 
   return (
@@ -20,6 +29,7 @@ const ChatInputContainer = () => {
         className="bg-[#B3B3B3] rounded-3xl px-3"
         onFocus={() => {}}
         onChange={(e) => onChangeInput(e)}
+        onEnterPress={handleSendMessage}
         textValue={inputVal || ''}
         placeholder="메시지를 입력하세요."
       />
