+++ src/app/_components/chat/containers/ChatPreviewContainer.tsx
@@ -1,0 +1,11 @@
+import ChatPreview from '../chatlist/ChatPreview'
+
+const ChatPreviewContainer = () => {
+  return (
+    <div>
+      <ChatPreview />
+    </div>
+  )
+}
+
+export default ChatPreviewContainer
